enhance type conversion handling in expression visitor; update parser to always parse integers as decimal; add comprehensive query builder tests

+++ tests/Dino.Core.Tests/DinoParserTests.cs
@@ -85,7 +85,7 @@ public class DinoParserTests
         whereExpr.Right.Should().BeOfType<DinoLiteralExpression>();
         var literal = (DinoLiteralExpression)whereExpr.Right;
         literal.Value.Should().Be(18);
-        literal.ValueType.Should().Be(typeof(int));
+        literal.ValueType.Should().Be(typeof(decimal));
     }
 
     [Fact]
